requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Authlib==1.1.0
black==22.8.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
discord.py==2.0.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-WTF==1.0.1
frozenlist==1.3.1
h11==0.13.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
outcome==1.2.0
pathspec==0.10.1
platformdirs==2.5.2
pycparser==2.21
PyNaCl==1.5.0
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
tomli==2.0.1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.12
waitress==2.1.2
webdriver-manager==3.8.3
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
yarl==1.8.1
